Temporary hack to workaround failure to compile long long support
* octave_to_python.cc (pytave::create_sint_array): Change template
parameter to "long" for now, does not compile with "long long".

+++ octave_to_python.cc
@@ -182,7 +182,7 @@ namespace pytave {
       // Numpy is kind enough to provide us with the old-style defines.
       else if (bytes == sizeof(long long)) {
          boost::integral_constant<bool, bytes == sizeof(long long) && bytes != sizeof(int)> inst;
-         return create_array<long long, CLASS>(value, PyArray_LONGLONG, inst);
+         return create_array<long, CLASS>(value, PyArray_LONGLONG, inst);
       }
       else if (bytes == sizeof(long)) {
          boost::integral_constant<bool, bytes == sizeof(long) && bytes != sizeof(int) && bytes != sizeof(long long)> inst;
